Split core into multiple files.
